refactor(core): Extract shared modules and utilities
- fetchTimer function moved to @hawk.so/core
- StackParser moved to @hawk.so/core
- event utility functions moved to @hawk.so/core
- selector utility functions moved to @hawk.so/core
- EventRejectedError moved to @hawk.so/core
