+++ packages/javascript/src/addons/breadcrumbs.ts
@@ -2,8 +2,7 @@
  * @file Breadcrumbs module - captures chronological trail of events before an error
  */
 import type { Breadcrumb, BreadcrumbLevel, BreadcrumbType, Json, JsonNode } from '@hawk.so/types';
-import { isValidBreadcrumb, log, Sanitizer } from '@hawk.so/core';
-import { buildElementSelector } from '../utils/selector';
+import { buildElementSelector, isValidBreadcrumb, log, Sanitizer } from '@hawk.so/core';
 
 /**
  * Default maximum number of breadcrumbs to store
